Fixed a bug with recentering at end of file
CCM doesn't recenter correctly when you have this option:

(setq ccm-recenter-at-end-of-file nil)

and do (end-of-buffer) or (ccm-scroll-up). It first recenters in
the middle of the window. Only then, after another command, it
recenters correctly at the bottom.

Line 341 doesn't work as intended because (window-end) doesn't
return the right value when the window scrolls. At that point in
time, when line 341 gets run, the value of (point) is bigger
then (window-end).

I don't know what kind of work line-end-position has to do behind
the scenes, but I haven't noticed any ill effects. But I also
haven't tested it with really big files.